test: move test_notification to helpers
