+++ test/helpers.h
@@ -4,6 +4,7 @@
 #include <glib.h>
 
 GVariant *notification_setup_raw_image(const char *path);
+struct notification *test_notification(const char *name, gint64 timeout);
 
 #endif
 /* vim: set tabstop=8 shiftwidth=8 expandtab textwidth=0: */
