+++ test/queues.h
@@ -31,25 +31,6 @@
 
 #define NOT_LAST(n) do {ASSERT_EQm("Notification " #n " should have been deleted.", 1, notification_refcount_get(n)); g_clear_pointer(&n, notification_unref); } while(0)
 
-static inline struct notification *test_notification(const char *name, gint64 timeout)
-{
-        struct notification *n = notification_create();
-
-        n->dbus_client = g_strconcat(":", name, NULL);
-        n->appname =     g_strconcat("app of ", name, NULL);
-        n->summary =     g_strconcat(name, NULL);
-        n->body =        g_strconcat("See, ", name, ", I've got a body for you!", NULL);
-
-        notification_init(n);
-
-        n->format = "%s\n%b";
-
-        if (timeout != -1)
-                n->timeout = S2US(timeout);
-
-        return n;
-}
-
 /* Retrieve a notification by its id. Solely for debugging purposes */
 struct notification *queues_debug_find_notification_by_id(int id);
